Add a "tx output spender" index
Adds an outpoint -> txid index, which can be used to find which transactions spent a given output.
This is extremely useful for Lightning and more generally for layer-2 protocol that rely on chains of unpublished transactions.
If enabled, this index will be used by `gettxspendingprevout` when it does not find a spending transaction in the mempool.