Relax MIN_STANDARD_TX_NONWITNESS_SIZE to 64 non-witness bytes exactly
Since the original fix was set to be a "reasonable" transaction
to reduce allocations and the true motivation later revealed,
it makes sense to relax this check to something more principled.

There are more exotic transaction patterns that could take advantage
of a relaxed requirement, such as 1 input, 1 output OP_RETURN to burn
a utxo to fees for CPFP purposes when change isn't practical.

Two changes could be accomplished:

1) Anything not 64 bytes could be allowed

2) Anything above 64 bytes could be allowed

To enable the maximum use-cases, (1) was taken.

The functional test is also modified to test the actual case
we care about: 64 bytes