bitcoin-wallet: make bitcoin-wallet tool load config file
Currently bitcoin-wallet accepts a -datadir argument but ignores and config
file in that directory. This changes it to load the config file for consistency
with other cli tools, and in case it contains any significant settings.

Also update various tests using the data directory to call ReadConfigFiles.

Motivation for these changes is to prevent errors after the next commit makes
it at assert error to get the datadir path without reading the configuration
first.