memusage: pool keeps track of its memory usage
That way all containers that use the pool have accurate memory tracking.

Add test to show memory is accurately tracked, even when nodes can't be allocated by the pool.
The more accurate memory estimation interferes shows that our memory estimation for Windows
is off, as the real allocated memory is much higher. This adapts the memusage_test test so it still
works with the more correct estimation.