Release `LockData::dd_mutex` before calling `*_detected` functions
Both `double_lock_detected()` and `potential_deadlock_detected()`
functions call `LogPrintf()` which in turn implies locking of the
`Logger::m_cs` mutex. To avoid a deadlock the latter must not have the
`Mutex` type.
With this change the mentioned restriction has been lifted, and it is
possible now to use our regular `Mutex` type for the `Logger::m_cs`
mutex instead of a dedicated `StdMutex` type.