refactor: PSBTError::MISSING_INPUTS
Rename PSBTError::MISSING_INPUTS to
PSBTError::INVALID_INPUTS

This error currently only covers the case where the specified outpoint
vout index number is not found in the input.

The change in wording provides more clarity in the codebase and to the
error, as an input may have been provided, but was invalid. Returning
"MISSING_INPUTS" can be ambiguous.