refactor: Reintroduce uint256S(const char*) as consteval
uint256S() calls taking C-string literals are littered throughout the code and were executed at runtime to perform parsing unless a given optimizer was surprisingly efficient. While this may not be a hot spot, it's better hygiene in C++20 to store the parsed data blob directly in the binary, without any parsing at runtime.

Verified that both GCC and Clang use the consteval overload of uint256S() for C-string literals by looking at produced binaries in objdump and confirming that calls to SetHex() do not remain in such a caller.