refactor: move merkle_root logic into KeyPair
Move the merkle_root tweaking logic into the `KeyPair` constructor. This
commit adds `merkle_root` to `CKey::ComputeKeyPair` and passes it to the
`KeyPair` constructor.

The `if (merkle_root)` block is move-only, except that
`keypair.data()` is changed to `m_keypair->data()`. This is because we
are no longer accessing `m_keypair` through the `KeyPair::data()` method
because we can now access the private member `m_keypair` directly.
Because of this, we can now make `KeyPair::data()` a read-only const
method.