[staging] crypto: add KeyPair wrapper class
Add a `KeyPair` class which wraps the `secp256k1_keypair`. This allows
keeps the secret data in secure memory and allows for passing the
`KeyPair` object directly to libsecp256k1 functions expecting a
`secp256k1_keypair`.

Motivation: when passing `CKeys` for taproot outputs to libsecp256k1 functions,
the first step is to create a `secp256k1_keypair` data type and use that
instead. This is so the libsecp256k1 function can determine if the key
needs to be negated, e.g., when signing.

This is a bit clunky in that it creates an extra step when using a `CKey`
for a taproot output and also involves copying the secret data into a
temporary object, which the caller must then take care to cleanse. In
addition, the logic for for applying the merkle_root tweak currently
only exists in the `SignSchnorr` function.

In a later commit, we will add the merkle_root tweaking logic to this
function, which will make the merkle_root logic reusable outside of
signing by using the `KeyPair` class directly.

Note: the `data()` method is not const in this commit but will be made
const in a later commit. The only reason for this is to keep the diffs
easier to analyze as we move the merkle_root logic out of `SignSchnorr`
and the finally move the `SignSchnorr` function from `CKey` to `KeyPair`.
This commit is labelled `[staging]` to make it clear that the class as
implemented with a non-const data method is not the desired state of the
class and that it will be resolved in a later commit.